fetcher/requirements.txt
websockets>=12.0
psycopg2-binary>=2.9
aiohttp>=3.9
